Change without git remote model
